add --force-section-merge flag
It seems that for relocatable links, ld does not perform string literal
merging. That is kind of understandable, since it would normally just be
wasted effort, as one doesn't lose any opportunities by postponing that
to the final link.

However, for VxWorks, all modules are built with -r, and there is no
final link - or rather, that is done by the runtime loader. By that
time, merging strings is too late to have any benefit; the memory is
already wasted. Since many VxWorks targets have rather limited disk and
RAM, saving even a few KB is worthwhile.

This is an attempt at adding a --force-section-merge command line flag,
similar to the --force-group-allocation flag. It is almost entirely
mechanical, and I fear that it can't be that easy (for example, there
might be some relocation fixup code hidden under a
!bfd_link_relocatable (&link_info)).